flask migrations with regorganisations all models are in on file thats is db models/modles
